OvmfPkg BasePciCapLib: Fix VS build failure
Fix VS warning C4244: 'function': conversion from 'UINT32' to 'UINT16',
possible loss of data.

Contributed-under: TianoCore Contribution Agreement 1.1
Cc: Laszlo Ersek <lersek@redhat.com>
[lersek@redhat.com: remove whitespace after casts]

+++ OvmfPkg/Library/BasePciCapLib/BasePciCapLib.c
@@ -613,8 +613,8 @@ PciCapListInit (
       }
 
       Status = InsertPciCap (OutCapList, CapHdrOffsets, PciCapExtended,
-                 ExtendedCapHdr.CapabilityId, ExtendedCapHdrOffset,
-                 ExtendedCapHdr.CapabilityVersion);
+                 (UINT16)ExtendedCapHdr.CapabilityId, ExtendedCapHdrOffset,
+                 (UINT8)ExtendedCapHdr.CapabilityVersion);
       if (RETURN_ERROR (Status)) {
         goto FreeCapHdrOffsets;
       }
